==> figure_AML_coevolution/R/00_AML_colors.R <==
AML.colors = c('HSC' = RColorBrewer::brewer.pal(name = 'YlOrRd', n = 5)[5],
               'Progenitor' = RColorBrewer::brewer.pal(name = 'YlOrRd', n = 5)[4],
               'Mono' = RColorBrewer::brewer.pal(name = 'YlOrRd', n = 5)[3],
               'DC' = RColorBrewer::brewer.pal(name = 'YlOrRd', n = 5)[2],
               'CD4+ T cell' = 'lightblue', 
               'CD8+ T cell' = 'darkblue', 
               'NK' = 'purple', 
               'B cell' = 'lightgreen', 
               'Plasma cell' = 'darkgreen')
